code rabbit 개선점 수정

+++ src/main/java/org/example/hugmeexp/domain/admin/service/AdminAttendanceService.java
@@ -45,8 +45,8 @@ public class AdminAttendanceService {
         double activeUserRate = totalUsers > 0 ? (double) activeUsers / totalUsers : 0.0;
 
         return ActiveUsersStatsResponse.builder()
-                .totalUsers((int) totalUsers)
-                .activeUsers((int) activeUsers)
+                .totalUsers(Math.toIntExact(totalUsers))
+                .activeUsers(Math.toIntExact(activeUsers))
                 .activeUserRate(activeUserRate)
                 .baseDate(today.toString())
                 .build();
